requirements.txt
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pandas
python-dotenv
pytz
requests
gspread
gspread-formatting
gspread-dataframe
numpy
